feat: move runtime dependent configs (repo and config) to runtime folder
