+++ package.json
@@ -8,9 +8,12 @@
   "main": "src/core/index.js",
   "browser": {
     "libp2p-ipfs-nodejs": "libp2p-ipfs-browser",
-    "./src/core/default-repo.js": "./src/core/default-repo-browser.js",
+    "./src/core/runtime/config-nodejs.js": "./src/core/runtime/config-browser.js",
+
+    "./src/core/runtime/libp2p-nodejs.js": "./src/core/runtime/libp2p-browser.js",
+    "./src/core/runtime/repo-nodejs.js": "./src/core/runtime/repo-browser.js",
     "./src/core/components/init-assets.js": false,
-    "./test/utils/create-repo-node.js": "./test/utils/create-repo-browser.js",
+    "./test/utils/create-repo-nodejs.js": "./test/utils/create-repo-browser.js",
     "stream": "readable-stream"
   },
   "engines": {
