+++ src/core/components/init.js
@@ -3,8 +3,8 @@
 const peerId = require('peer-id')
 const waterfall = require('async/waterfall')
 const parallel = require('async/parallel')
-const isNode = require('detect-node')
 const promisify = require('promisify-es6')
+const config = require('../runtime/config-nodejs.json')
 
 const addDefaultAssets = require('./init-assets')
 
@@ -36,10 +36,6 @@ module.exports = function init (self) {
     opts.emptyRepo = opts.emptyRepo || false
     opts.bits = Number(opts.bits) || 2048
     opts.log = opts.log || function () {}
-
-    const config = isNode
-      ? require('../../init-files/default-config-node.json')
-      : require('../../init-files/default-config-browser.json')
 
     waterfall([
       // Verify repo does not yet exist.
